[Offline]: speedup example github action
