+++ .github/workflows/example.yml
@@ -8,7 +8,34 @@ on:
     types: [published]
 
 jobs:
-  build:
+  example:
+    runs-on: ubuntu-22.04
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          submodules: true
+      - name: Prepare Local Xmake
+        run: cp -rf . ../xmake-source
+      - uses: xmake-io/github-action-setup-xmake@v1
+        with:
+          xmake-version: local#../xmake-source
+      - name: Prepare Environment
+        run: |
+          sudo apt update
+          sudo apt install -y build-essential lcov
+          sudo apt install -y ccache
+          sudo apt install -y protobuf-compiler
+          sudo apt install -y libgflags-dev
+          sudo apt install -y g++-aarch64-linux-gnu
+      - name: Run Examples
+        env:
+          XMAKE_ROOT: y
+          XMAKE_MAIN_REPO: https://github.com/zxmake/zxmake-repo.git
+          XMAKE_BINARY_REPO: https://github.com/zxmake/zxmake-build-artifacts.git
+        run: |
+          stdbuf -oL bash scripts/example.sh
+
+  cross-example:
     runs-on: ubuntu-22.04
     steps:
       - uses: actions/checkout@v2
@@ -25,12 +52,10 @@ jobs:
           sudo apt install -y build-essential lcov
           sudo apt install -y ccache
           sudo apt install -y g++-aarch64-linux-gnu
-          sudo apt install -y protobuf-compiler
-          sudo apt install -y libgflags-dev
-      - name: Run Examples
+      - name: Run Cross-Build Examples
         env:
           XMAKE_ROOT: y
           XMAKE_MAIN_REPO: https://github.com/zxmake/zxmake-repo.git
           XMAKE_BINARY_REPO: https://github.com/zxmake/zxmake-build-artifacts.git
         run: |
-          bash scripts/example.sh
+          stdbuf -oL bash scripts/example-cross.sh
