+++ .github/workflows/freebsd.yml
@@ -1,6 +1,7 @@
 name: FreeBSD
 
 on:
+  workflow_dispatch:
   # pull_request:
   # push:
   #   branches: [ master ]
