+++ .github/workflows/macos.yml
@@ -1,6 +1,7 @@
 name: macOS (x86_64)
 
 on:
+  workflow_dispatch:
   # pull_request:
   # push:
   #   branches: [ master ]
