--- example/cross-build/aarch64-none-linux-gnu/pb/header.proto
@@ -1,9 +1,0 @@
-syntax = "proto3";
-
-package github.tomocat.proto;
-
-message End2EndSource {
-  string module_name = 1;
-  uint32 sequence_num = 2;
-  double transit_time_ms = 3;
-}
