+++ example/cross-build/android/src/main.cc
@@ -1,6 +1,6 @@
 #include <iostream>
 
-int main(int argc, char** argv) {
+int main() {
     std::cout << "hello world!" << std::endl;
     return 0;
 }
