+++ example/cross-build/arm-none-linux-gnueabihf/test.sh
@@ -28,7 +28,7 @@ else
         warning "file [${SDK_ZIP_FILE}] exists. skip downloading"
     else
         info "start downloading ndk zip file [${SDK_ZIP_FILE}]"
-        wget -q "https://developer.arm.com/-/media/Files/downloads/gnu/13.3.rel1/binrel/arm-gnu-toolchain-13.3.rel1-x86_64-arm-none-linux-gnueabihf.tar.xz" -O ${SDK_ZIP_FILE}
+        wget "https://github.com/TOMO-CAT/zxmake-mirror/releases/download/20260510.1/arm-gnu-toolchain-13.3.rel1-x86_64-arm-none-linux-gnueabihf.tar.xz" -O ${SDK_ZIP_FILE}
     fi
     tar -xf ${SDK_ZIP_FILE}
 fi
