+++ example/cross-build/arm-none-linux-gnueabihf/xmake.lua
@@ -3,13 +3,18 @@ add_rules("mode.debug", "mode.release")
 set_languages("c++17")
 
 -- 编译出 host 版本的 protoc, 用于生成 *.pb.h 和 *.pb.cc
-add_requires("protobuf-cpp~host 3.19.4", {host = true, alias = "protoc"})
+-- * 方案 1: 源码编译, 指定 host = true 就只会编译 host 版本的 protoc
+-- add_requires("protobuf-cpp~host 3.19.4", {host = true, alias = "protoc"})
+-- * 方案 2: 直接使用 protoc 预编译好的二进制, 节省编译速度
+add_requires("protoc 3.19.4")
+
 -- 编译出 cross 版本的 protobuf 库, 用于构造交叉编译成品库
-add_requires("protobuf-cpp 3.19.4")
+-- `protoc = false` 会跳过编译 protoc, 因为交叉编译时本身就只需要 host protoc, 可以优化编译速度
+add_requires("protobuf-cpp 3.19.4", {configs = {protoc = false}})
 
 target("pb", function()
     set_kind("object")
-    add_files("pb/*.proto", { proto_public = true})
+    add_files("pb/*.proto", {proto_public = true})
     add_rules("protobuf.cpp")
     add_packages("protobuf-cpp", {public = true})
